@@ -1,8 +1,14 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+import 'package:clima_core/either.dart';
 import 'package:clima_core/failure.dart';
 import 'package:clima_data/models/city_model.dart';
 import 'package:clima_data/providers.dart';
 import 'package:clima_domain/entities/city.dart';
-import 'package:dartz/dartz.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -27,4 +33,5 @@ class CityLocalDataSource {
 }
 
 final cityLocalDataSourceProvider = Provider(
-    (ref) => CityLocalDataSource(ref.watch(sharedPreferencesProvider)));
+  (ref) => CityLocalDataSource(ref.watch(sharedPreferencesProvider)),
+);
