@@ -1,89 +1,114 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
 import 'package:clima_domain/entities/city.dart';
-import 'package:clima_ui/screens/settings_screen.dart';
+import 'package:clima_domain/entities/unit_system.dart';
+import 'package:clima_ui/state_notifiers/api_key_state_notifier.dart' as a;
 import 'package:clima_ui/state_notifiers/city_state_notifier.dart' as c;
-import 'package:clima_ui/state_notifiers/forecasts_state_notifier.dart' as f;
-import 'package:clima_ui/state_notifiers/weather_state_notifier.dart' as w;
+import 'package:clima_ui/state_notifiers/full_weather_state_notifier.dart' as w;
+import 'package:clima_ui/state_notifiers/unit_system_state_notifier.dart' as u;
 import 'package:clima_ui/utilities/constants.dart';
-import 'package:clima_ui/utilities/failure_snack_bar.dart';
 import 'package:clima_ui/utilities/hooks.dart';
-import 'package:clima_ui/widgets/bottom_row.dart';
-import 'package:clima_ui/widgets/current_conditions.dart';
-import 'package:clima_ui/widgets/forecast_widget.dart';
-import 'package:clima_ui/widgets/reusable_widgets.dart';
+import 'package:clima_ui/utilities/snack_bars.dart';
+import 'package:clima_ui/widgets/others/failure_banner.dart';
+import 'package:clima_ui/widgets/others/overflow_menu_button.dart';
+import 'package:clima_ui/widgets/weather/additional_info_widget.dart';
+import 'package:clima_ui/widgets/weather/daily_forecasts_widget.dart';
+import 'package:clima_ui/widgets/weather/hourly_forecasts_widget.dart';
+import 'package:clima_ui/widgets/weather/main_info_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
 import 'package:material_floating_search_bar/material_floating_search_bar.dart';
 import 'package:sizer/sizer.dart';
-import 'package:url_launcher/url_launcher.dart';
 
-enum Menu { settings, help }
-
-class WeatherScreen extends HookWidget {
+class WeatherScreen extends HookConsumerWidget {
   const WeatherScreen({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
-    final weatherState = useProvider(w.weatherStateNotifierProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    final fullWeatherState = ref.watch(w.fullWeatherStateNotifierProvider);
+    final fullWeather = fullWeatherState.fullWeather;
 
-    final forecastsStateNotifier =
-        useProvider(f.forecastsStateNotifierProvider.notifier);
+    final fullWeatherStateNotifier =
+        ref.watch(w.fullWeatherStateNotifierProvider.notifier);
 
-    final weatherStateNotifier =
-        useProvider(w.weatherStateNotifierProvider.notifier);
+    final apiKeyStateNotifier =
+        ref.watch(a.apiKeyStateNotifierProvider.notifier);
+
     final controller = useFloatingSearchBarController();
 
-    final isLoading = useState(weatherState is c.Loading);
+    final cityStateNotifier = ref.watch(c.cityStateNotifierProvider.notifier);
+    final cityState = ref.watch(c.cityStateNotifierProvider);
 
-    final cityStateNotifier = useProvider(c.cityStateNotifierProvider.notifier);
+    final refreshIndicatorKey = useGlobalKey<RefreshIndicatorState>();
 
-    Future<void> load() async {
-      isLoading.value = true;
-      await Future.wait(
-        [
-          weatherStateNotifier.loadWeather(),
-          forecastsStateNotifier.loadForecasts(),
-        ],
-      );
-      isLoading.value = false;
-    }
+    final unitSystemStateNotifier =
+        ref.watch(u.unitSystemStateNotifierProvider.notifier);
 
-    useEffect(
-      () => cityStateNotifier.addListener((state) {
-        if (state is c.Error) {
-          showFailureSnackBar(context, failure: state.failure, duration: 2);
-        }
-      }),
-      [cityStateNotifier],
+    final unitSystem = ref.watch(
+      u.unitSystemStateNotifierProvider.select((state) => state.unitSystem),
     );
 
     useEffect(
-      () => weatherStateNotifier.addListener((state) {
-        if (state is w.Error) {
-          showFailureSnackBar(context, failure: state.failure, duration: 2);
-        }
-      }),
-      [weatherStateNotifier],
+      () {
+        Future.microtask(
+          () => Future.wait([
+            apiKeyStateNotifier.loadApiKey(),
+            unitSystemStateNotifier.loadUnitSystem(),
+            fullWeatherStateNotifier.loadFullWeather(),
+          ]),
+        );
+
+        return null;
+      },
+      [fullWeatherStateNotifier, unitSystemStateNotifier, apiKeyStateNotifier],
     );
 
     useEffect(
-      () => forecastsStateNotifier.addListener((state) {
-        if (state is f.Error) {
-          showFailureSnackBar(context, failure: state.failure, duration: 2);
+      () {
+        if (fullWeather == null) {
+          return null;
         }
-      }),
-      [forecastsStateNotifier],
+
+        return cityStateNotifier.addListener((state) {
+          if (state is c.Error) {
+            showFailureSnackBar(context, failure: state.failure, duration: 2);
+          }
+        });
+      },
+      [cityStateNotifier, fullWeather == null],
+    );
+
+    useEffect(
+      () {
+        if (fullWeather == null) {
+          return null;
+        }
+
+        return fullWeatherStateNotifier.addListener((state) {
+          if (state is w.Error) {
+            showFailureSnackBar(context, failure: state.failure, duration: 2);
+          }
+        });
+      },
+      [fullWeatherStateNotifier, fullWeather],
     );
 
     return Scaffold(
       resizeToAvoidBottomInset: false,
       body: FloatingSearchAppBar(
+        liftOnScrollElevation: 0.0,
+        elevation: fullWeather == null ? 2.0 : 0.0,
         systemOverlayStyle: Theme.of(context).appBarTheme.systemOverlayStyle,
         automaticallyImplyBackButton: false,
         controller: controller,
-        progress: isLoading.value,
+        progress: fullWeatherState is w.Loading || cityState is c.Loading,
+        accentColor: Theme.of(context).colorScheme.secondary,
         onSubmitted: (String newCityName) async {
           controller.close();
 
@@ -92,40 +117,54 @@ class WeatherScreen extends HookWidget {
             return;
           }
 
-          isLoading.value = true;
           await cityStateNotifier.setCity(City(name: trimmedCityName));
-          if (context.read(c.cityStateNotifierProvider) is! c.Error) {
-            await Future.wait([
-              weatherStateNotifier.loadWeather(),
-              forecastsStateNotifier.loadForecasts(),
-            ]);
+          if (ref.read(c.cityStateNotifierProvider) is! c.Error) {
+            await fullWeatherStateNotifier.loadFullWeather();
           }
-          isLoading.value = false;
         },
-        title: Text(
-          DateFormat('EEEE, d MMMM yyyy').format(DateTime.now()),
-          style: TextStyle(
-            color: Theme.of(context).textTheme.subtitle2!.color,
-            fontSize:
-                MediaQuery.of(context).size.shortestSide < kTabletBreakpoint
-                    ? 11.sp
-                    : 8.sp,
-          ),
-        ),
+        title: fullWeather == null
+            // If we use `null` here, it would show the hint. We want it to
+            // show nothing.
+            ? const SizedBox.shrink()
+            : Row(
+                children: [
+                  Text(
+                    'Updated ${DateFormat.Md().add_jm().format(fullWeather.currentWeather.date.toLocal())} · ',
+                    style: TextStyle(
+                      color: Theme.of(context).textTheme.subtitle2!.color,
+                      fontSize: MediaQuery.of(context).size.shortestSide <
+                              kTabletBreakpoint
+                          ? 11.sp
+                          : 5.sp,
+                    ),
+                  ),
+                  Text(
+                    () {
+                      switch (unitSystem) {
+                        case UnitSystem.metric:
+                          return '°C';
+
+                        case UnitSystem.imperial:
+                          return '°F';
+
+                        case null:
+                          return '';
+                      }
+                    }(),
+                    style: TextStyle(
+                      color: Theme.of(context).textTheme.subtitle1!.color,
+                      fontSize: MediaQuery.of(context).size.shortestSide <
+                              kTabletBreakpoint
+                          ? 11.sp
+                          : 5.sp,
+                    ),
+                  ),
+                ],
+              ),
         hint: 'Enter city name',
-        color: Theme.of(context).appBarTheme.color,
+        color: Theme.of(context).appBarTheme.backgroundColor,
         transitionCurve: Curves.easeInOut,
         leadingActions: [
-          FloatingSearchBarAction(
-            child: CircularButton(
-              icon: Icon(
-                Icons.refresh,
-                color: Theme.of(context).appBarTheme.iconTheme!.color,
-              ),
-              tooltip: 'Refresh',
-              onPressed: load,
-            ),
-          ),
           FloatingSearchBarAction.back(
             color: Theme.of(context).appBarTheme.iconTheme!.color,
           ),
@@ -143,103 +182,86 @@ class WeatherScreen extends HookWidget {
               },
             ),
           ),
-          FloatingSearchBarAction(
-            child: PopupMenuButton(
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(5.0),
-              ),
-              offset: const Offset(512.0, -512.0),
-              icon: Icon(
-                Icons.more_vert,
-                color: Theme.of(context).appBarTheme.iconTheme!.color,
-              ),
-              tooltip: 'More options',
-              itemBuilder: (BuildContext context) => <PopupMenuEntry<Menu>>[
-                PopupMenuItem<Menu>(
-                  value: Menu.settings,
-                  child: ListTile(
-                    title: const Text('Settings'),
-                    onTap: () {
-                      Navigator.of(context).pop();
-                      Navigator.push(
-                        context,
-                        MaterialPageRoute(
-                          builder: (BuildContext context) => SettingScreen(),
-                        ),
-                      );
-                    },
-                  ),
-                ),
-                PopupMenuItem<Menu>(
-                  value: Menu.help,
-                  child: ListTile(
-                    title: const Text('Help & feedback'),
-                    onTap: () {
-                      Navigator.of(context).pop();
-                      showGeneralSheet(
-                        context,
-                        title: 'Help & feedback',
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            SettingsTile(
-                              title: 'Submit issue',
-                              leading: Icon(
-                                Icons.bug_report_outlined,
-                                color: Theme.of(context).iconTheme.color,
-                              ),
-                              onTap: () => launch(
-                                'https://github.com/CentaurusApps/clima/issues/new',
-                              ),
-                            ),
-                            SettingsTile(
-                              title: 'Contact developer',
-                              leading: Icon(
-                                Icons.email_outlined,
-                                color: Theme.of(context).iconTheme.color,
-                              ),
-                              onTap: () => launch(
-                                'mailto:apps_centaurus@protonmail.com',
-                              ),
-                            ),
-                          ],
-                        ),
-                      );
-                    },
-                  ),
-                ),
-              ],
-            ),
-          ),
+          const FloatingSearchBarAction(child: OverflowMenuButton()),
           FloatingSearchBarAction.searchToClear(
             color: Theme.of(context).appBarTheme.iconTheme!.color,
             showIfClosed: false,
           )
         ],
-        body: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Flexible(
-              flex: 10,
-              child: CurrentConditions(),
+        body: SafeArea(
+          child: RefreshIndicator(
+            key: refreshIndicatorKey,
+            onRefresh: fullWeatherStateNotifier.loadFullWeather,
+            color: Theme.of(context).textTheme.subtitle1!.color,
+            child: Padding(
+              padding: EdgeInsets.symmetric(
+                horizontal: () {
+                  if (MediaQuery.of(context).orientation ==
+                          Orientation.landscape &&
+                      MediaQuery.of(context).size.shortestSide >
+                          kTabletBreakpoint) {
+                    return 10.0.w;
+                  } else if (MediaQuery.of(context).orientation ==
+                          Orientation.landscape &&
+                      MediaQuery.of(context).size.shortestSide <
+                          kTabletBreakpoint) {
+                    return 35.0.w;
+                  } else {
+                    return 5.0.w;
+                  }
+                }(),
+              ),
+              child: fullWeatherState is w.Error && fullWeather == null
+                  ? FailureBanner(
+                      failure: fullWeatherState.failure,
+                      onRetry: fullWeatherStateNotifier.loadFullWeather,
+                    )
+                  : fullWeather == null
+                      ? null
+                      : Container(
+                          constraints: const BoxConstraints.expand(),
+                          child: SingleChildScrollView(
+                            physics: const BouncingScrollPhysics(),
+                            child: Column(
+                              children: [
+                                const MainInfoWidget(),
+                                Divider(
+                                  color: Theme.of(context)
+                                      .textTheme
+                                      .subtitle1!
+                                      .color!
+                                      .withAlpha(65),
+                                ),
+                                SizedBox(
+                                  height:
+                                      MediaQuery.of(context).size.aspectRatio ==
+                                              1.0
+                                          ? 24.h
+                                          : 16.h,
+                                  child: const HourlyForecastsWidget(),
+                                ),
+                                Divider(
+                                  color: Theme.of(context)
+                                      .textTheme
+                                      .subtitle1!
+                                      .color!
+                                      .withAlpha(65),
+                                ),
+                                const DailyForecastsWidget(),
+                                Divider(
+                                  color: Theme.of(context)
+                                      .textTheme
+                                      .subtitle1!
+                                      .color!
+                                      .withAlpha(65),
+                                ),
+                                const AdditionalInfoWidget(),
+                              ],
+                            ),
+                          ),
+                        ),
             ),
-            Divider(
-              color:
-                  Theme.of(context).textTheme.subtitle1!.color!.withAlpha(65),
-            ),
-            Flexible(
-              flex: 2,
-              child: ForecastHorizontal(),
-            ),
-            Divider(
-              color:
-                  Theme.of(context).textTheme.subtitle1!.color!.withAlpha(65),
-            ),
-            Flexible(
-              flex: 2,
-              child: BottomRow(),
-            ),
-          ],
+          ),
         ),
       ),
     );
